+++ js_utils/text_validation.js
@@ -1,5 +1,72 @@
 "use strict";
 
+const _validate_old_barcode = function (barcode) {
+  const plate_barcode_len = barcode.length;
+  // check that barcode is a valid length
+  if (plate_barcode_len < 10 || plate_barcode_len >= 12) {
+    return " ";
+  }
+  const start_code = barcode.slice(0, 2); // Tanner (7/30/21): "old" barcodes with the same format are MA, MB, ME
+  if (!["MA", "MB", "ME"].includes(start_code)) {
+    return " ";
+  }
+  for (let i = 2; i < plate_barcode_len; i++) {
+    const scan_ascii = barcode.charCodeAt(i);
+    // check that remaining characters are numeric
+    if (scan_ascii < 47 || scan_ascii > 58) {
+      return " ";
+    }
+  }
+  // check year is in range 00 to 99   [2 characters]
+  const year_code = barcode.slice(2, 4);
+  const year = parseInt(year_code);
+  if (year < 0 || year > 99) {
+    return " ";
+  }
+  // check day code is of range 001 to 366 [3 characters]
+  const day_code = barcode.slice(4, 7);
+  const day = parseInt(day_code);
+  if (day < 1 || day > 366) {
+    return " ";
+  }
+  return "";
+};
+
+const _validate_new_barcode = function (barcode) {
+  const plate_barcode_len = barcode.length;
+  // check that barcode is a valid length
+  if (plate_barcode_len !== 12) {
+    return " ";
+  }
+  // Tanner (7/30/21): Assuming only barcodes with header 'ML' are passed to this function, so not checking header
+  for (let i = 2; i < plate_barcode_len; i++) {
+    const scan_ascii = barcode.charCodeAt(i);
+    // check that remaining characters are numeric
+    if (scan_ascii < 47 || scan_ascii > 58) {
+      return " ";
+    }
+  }
+  // check year is at least 2021   [2 characters]
+  const year_code = barcode.slice(2, 6);
+  const year = parseInt(year_code);
+  if (year < 2021) {
+    return " ";
+  }
+  // check day code is of range 001 to 366 [3 characters]
+  const day_code = barcode.slice(6, 9);
+  const day = parseInt(day_code);
+  if (day < 1 || day > 366) {
+    return " ";
+  }
+  // check kit ID
+  const kit_id_remainder = parseInt(barcode.slice(9)) % 4;
+  // Tanner (7/30/21): currently this number must have a remainder of 0 or 1 when divided by 4
+  if (kit_id_remainder > 1) {
+    return " ";
+  }
+  return "";
+};
+
 // dependencies
 const uuidBase62 = require("@tofandel/uuid-base62"); // External library depenency of @tofandel/uuid-base62
 /* eslint-disable new-cap */
@@ -60,64 +127,19 @@ export class TextValidation {
   /**
    * Returns the feedback text for the plate barcode validation
    *
-   * @param  {text}   text The text on which the validation rules are verified
+   * @param  {text}   barcode The text on which the validation rules are verified
    * @return {string}  The string is either empty on valid and <space> on invalid
    *
    */
-  validate_plate_barcode(text) {
-    // refactored for 100% code coverage on JEST Unit testcases.
-    let response = " "; // let us first assume that incoming text is invalid unless it passes the criteria.
-    // As the validation rules are on Vuex data of barcode and this can be null a check is required.
-    if (text == null) {
-      return response;
+  validate_plate_barcode(barcode) {
+    // TODOs tanner
+    // - make sure tests with old barcodes still pass
+    // - add tests for new barcodes
+    // - make sure to test ""
+    if (barcode == null) {
+      return " ";
     }
-    const platebarcode_len = text.length;
-    if (platebarcode_len < 10) {
-      // clearly there is no need to process as its not matching basic rule length (< 10)
-      return response;
-    }
-    if (platebarcode_len >= 12) {
-      // clearly there is no need to process as its not matching basic rule length (> 11)
-      return response;
-    }
-    const start_code = text.slice(0, 2); // this has to be MA, MB or MD [2 characters]
-    if (start_code.charAt(0) != "M") {
-      // clearly the first criteria of two characters is not matching
-      return response;
-    }
-    if (start_code.charAt(1) != "A") {
-      if (start_code.charAt(1) != "B") {
-        if (start_code.charAt(1) != "E") {
-          return response;
-        }
-      }
-    }
-    for (let i = 2; i < platebarcode_len; i++) {
-      const scan_ascii = text.charCodeAt(i);
-      if (scan_ascii < 47) {
-        // clearly the valid input from second character in not numeral [0 to 9] ascii 0 -> 47
-        return response;
-      }
-      if (scan_ascii > 58) {
-        // clearly the valid input from second charcter is not numeral [0 to 9] ascii 9 -> 57
-        return response;
-      }
-    }
-    const year_code = text.slice(2, 4); // this is of range 00 to 99   [2 characters]
-    const year = parseInt(year_code);
-    if (year >= 0 && year <= 99) {
-      response = "";
-    }
-
-    const day_code = text.slice(4, 7); // this is of range 001 to 366 [3 characters]
-    const day = parseInt(day_code);
-
-    if (day > 0 && day < 367) {
-      response = "";
-    } else {
-      response = " ";
-    }
-    return response;
+    return barcode.slice(0, 2) === "ML" ? _validate_new_barcode(barcode) : _validate_old_barcode(barcode);
   }
   /**
    * Returns the feedback text for the uuidBase57 encoding validation